+++ packages/core/src/domain/session/oldCookiesMigration.spec.ts
@@ -6,15 +6,15 @@ import {
   tryOldCookiesMigration,
 } from './oldCookiesMigration'
 import { SESSION_EXPIRATION_DELAY } from './sessionConstants'
-import { SESSION_COOKIE_NAME, initCookieStore } from './sessionCookieStore'
+import { SESSION_COOKIE_NAME, initCookieStrategy } from './storeStrategies/sessionInCookie'
 
 describe('old cookies migration', () => {
-  const sessionStore = initCookieStore({ clientToken: '123' })!
+  const sessionStoreStrategy = initCookieStrategy({ clientToken: '123' })!
 
   it('should not touch current cookie', () => {
     setCookie(SESSION_COOKIE_NAME, 'id=abcde&rum=0&logs=1&expire=1234567890', SESSION_EXPIRATION_DELAY)
 
-    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStore)
+    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStoreStrategy)
 
     expect(getCookie(SESSION_COOKIE_NAME)).toBe('id=abcde&rum=0&logs=1&expire=1234567890')
   })
@@ -24,7 +24,7 @@ describe('old cookies migration', () => {
     setCookie(OLD_LOGS_COOKIE_NAME, '1', SESSION_EXPIRATION_DELAY)
     setCookie(OLD_RUM_COOKIE_NAME, '0', SESSION_EXPIRATION_DELAY)
 
-    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStore)
+    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStoreStrategy)
 
     expect(getCookie(SESSION_COOKIE_NAME)).toContain('id=abcde')
     expect(getCookie(SESSION_COOKIE_NAME)).toContain('rum=0')
@@ -35,7 +35,7 @@ describe('old cookies migration', () => {
   it('should create new cookie from a single old cookie', () => {
     setCookie(OLD_RUM_COOKIE_NAME, '0', SESSION_EXPIRATION_DELAY)
 
-    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStore)
+    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStoreStrategy)
 
     expect(getCookie(SESSION_COOKIE_NAME)).not.toContain('id=')
     expect(getCookie(SESSION_COOKIE_NAME)).toContain('rum=0')
@@ -43,7 +43,7 @@ describe('old cookies migration', () => {
   })
 
   it('should not create a new cookie if no old cookie is present', () => {
-    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStore)
+    tryOldCookiesMigration(SESSION_COOKIE_NAME, sessionStoreStrategy)
     expect(getCookie(SESSION_COOKIE_NAME)).toBeUndefined()
   })
 })
