+++ packages/core/src/domain/session/sessionManager.ts
@@ -6,8 +6,8 @@ import { relativeNow, clocksOrigin, ONE_MINUTE } from '../../tools/utils/timeUti
 import { DOM_EVENT, addEventListener, addEventListeners } from '../../browser/addEventListener'
 import { clearInterval, setInterval } from '../../tools/timer'
 import { SESSION_TIME_OUT_DELAY } from './sessionConstants'
-import { startSessionStoreManager } from './sessionStoreManager'
-import type { SessionStore } from './sessionStore'
+import { startSessionStore } from './sessionStore'
+import type { SessionStoreStrategy } from './storeStrategies/sessionStoreStrategy'
 
 export interface SessionManager<TrackingType extends string> {
   findActiveSession: (startTime?: RelativeTime) => SessionContext<TrackingType> | undefined
@@ -26,41 +26,41 @@ const SESSION_CONTEXT_TIMEOUT_DELAY = SESSION_TIME_OUT_DELAY
 let stopCallbacks: Array<() => void> = []
 
 export function startSessionManager<TrackingType extends string>(
-  sessionStore: SessionStore,
+  sessionStoreStrategy: SessionStoreStrategy,
   productKey: string,
   computeSessionState: (rawTrackingType?: string) => { trackingType: TrackingType; isTracked: boolean }
 ): SessionManager<TrackingType> {
-  const sessionStoreManager = startSessionStoreManager(sessionStore, productKey, computeSessionState)
-  stopCallbacks.push(() => sessionStoreManager.stop())
+  const sessionStore = startSessionStore(sessionStoreStrategy, productKey, computeSessionState)
+  stopCallbacks.push(() => sessionStore.stop())
 
   const sessionContextHistory = new ValueHistory<SessionContext<TrackingType>>(SESSION_CONTEXT_TIMEOUT_DELAY)
   stopCallbacks.push(() => sessionContextHistory.stop())
 
-  sessionStoreManager.renewObservable.subscribe(() => {
+  sessionStore.renewObservable.subscribe(() => {
     sessionContextHistory.add(buildSessionContext(), relativeNow())
   })
-  sessionStoreManager.expireObservable.subscribe(() => {
+  sessionStore.expireObservable.subscribe(() => {
     sessionContextHistory.closeActive(relativeNow())
   })
 
-  sessionStoreManager.expandOrRenewSession()
+  sessionStore.expandOrRenewSession()
   sessionContextHistory.add(buildSessionContext(), clocksOrigin().relative)
 
-  trackActivity(() => sessionStoreManager.expandOrRenewSession())
-  trackVisibility(() => sessionStoreManager.expandSession())
+  trackActivity(() => sessionStore.expandOrRenewSession())
+  trackVisibility(() => sessionStore.expandSession())
 
   function buildSessionContext() {
     return {
-      id: sessionStoreManager.getSession().id!,
-      trackingType: sessionStoreManager.getSession()[productKey] as TrackingType,
+      id: sessionStore.getSession().id!,
+      trackingType: sessionStore.getSession()[productKey] as TrackingType,
     }
   }
 
   return {
     findActiveSession: (startTime) => sessionContextHistory.find(startTime),
-    renewObservable: sessionStoreManager.renewObservable,
-    expireObservable: sessionStoreManager.expireObservable,
-    expire: sessionStoreManager.expire,
+    renewObservable: sessionStore.renewObservable,
+    expireObservable: sessionStore.expireObservable,
+    expire: sessionStore.expire,
   }
 }
