+++ packages/core/src/domain/session/sessionStoreOperations.spec.ts
@@ -2,10 +2,10 @@ import type { StubStorage } from '../../../test'
 import { mockClock, stubCookieProvider, stubLocalStorageProvider } from '../../../test'
 import type { InitConfiguration } from '../configuration'
 import { SESSION_EXPIRATION_DELAY } from './sessionConstants'
-import { initCookieStore, SESSION_COOKIE_NAME } from './sessionCookieStore'
-import { initLocalStorage, LOCAL_STORAGE_KEY } from './sessionLocalStorageStore'
-import type { SessionState } from './sessionStore'
-import { toSessionString } from './sessionStore'
+import { initCookieStrategy, SESSION_COOKIE_NAME } from './storeStrategies/sessionInCookie'
+import { initLocalStorageStrategy, LOCAL_STORAGE_KEY } from './storeStrategies/sessionInLocalStorage'
+import type { SessionState } from './sessionState'
+import { toSessionString } from './sessionState'
 import {
   processSessionStoreOperations,
   isLockEnabled,
@@ -19,18 +19,18 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
   [
     {
       title: 'Cookie Storage',
-      sessionStore: initCookieStore(initConfiguration)!,
+      sessionStoreStrategy: initCookieStrategy(initConfiguration)!,
       stubStorageProvider: stubCookieProvider,
       storageKey: SESSION_COOKIE_NAME,
     },
     {
       title: 'Local Storage',
-      sessionStore: initLocalStorage()!,
+      sessionStoreStrategy: initLocalStorageStrategy()!,
       stubStorageProvider: stubLocalStorageProvider,
       storageKey: LOCAL_STORAGE_KEY,
     },
   ] as const
-).forEach(({ title, sessionStore, stubStorageProvider, storageKey }) => {
+).forEach(({ title, sessionStoreStrategy, stubStorageProvider, storageKey }) => {
   describe(`process operations mechanism with ${title}`, () => {
     let initialSession: SessionState
     let otherSession: SessionState
@@ -39,7 +39,7 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
     let stubStorage: StubStorage
 
     beforeEach(() => {
-      sessionStore.clearSession()
+      sessionStoreStrategy.clearSession()
       initialSession = { id: '123', created: '0' }
       otherSession = { id: '456', created: '100' }
       processSpy = jasmine.createSpy('process')
@@ -53,49 +53,49 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
       })
 
       it('should persist session when process returns a value', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue({ ...otherSession })
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith(initialSession)
         const expectedSession = { ...otherSession, expire: jasmine.any(String) }
-        expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
         expect(afterSpy).toHaveBeenCalledWith(expectedSession)
       })
 
       it('should clear session when process returns an empty value', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue({})
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith(initialSession)
         const expectedSession = {}
-        expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
         expect(afterSpy).toHaveBeenCalledWith(expectedSession)
       })
 
       it('should not persist session when process returns undefined', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue(undefined)
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith(initialSession)
-        expect(sessionStore.retrieveSession()).toEqual(initialSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(initialSession)
         expect(afterSpy).toHaveBeenCalledWith(initialSession)
       })
 
       it('LOCK_MAX_TRIES value should not influence the behavior when lock mechanism is not enabled', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue({ ...otherSession })
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore, LOCK_MAX_TRIES)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy, LOCK_MAX_TRIES)
 
         expect(processSpy).toHaveBeenCalledWith(initialSession)
         const expectedSession = { ...otherSession, expire: jasmine.any(String) }
-        expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
         expect(afterSpy).toHaveBeenCalledWith(expectedSession)
       })
     })
@@ -106,37 +106,37 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
       })
 
       it('should persist session when process returns a value', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.callFake((session) => ({ ...otherSession, lock: session.lock }))
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith({ ...initialSession, lock: jasmine.any(String) })
         const expectedSession = { ...otherSession, expire: jasmine.any(String) }
-        expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
         expect(afterSpy).toHaveBeenCalledWith(expectedSession)
       })
 
       it('should clear session when process returns an empty value', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue({})
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith({ ...initialSession, lock: jasmine.any(String) })
         const expectedSession = {}
-        expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
         expect(afterSpy).toHaveBeenCalledWith(expectedSession)
       })
 
       it('should not persist session when process returns undefined', () => {
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         processSpy.and.returnValue(undefined)
 
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         expect(processSpy).toHaveBeenCalledWith({ ...initialSession, lock: jasmine.any(String) })
-        expect(sessionStore.retrieveSession()).toEqual(initialSession)
+        expect(sessionStoreStrategy.retrieveSession()).toEqual(initialSession)
         expect(afterSpy).toHaveBeenCalledWith(initialSession)
       })
 
@@ -195,7 +195,7 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
             }),
           })
           initialSession.expire = String(Date.now() + SESSION_EXPIRATION_DELAY)
-          sessionStore.persistSession(initialSession)
+          sessionStoreStrategy.persistSession(initialSession)
           processSpy.and.callFake((session) => ({ ...session, processed: 'processed' } as SessionState))
 
           processSessionStoreOperations(
@@ -217,12 +217,12 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
                   processed: 'processed',
                   expire: jasmine.any(String),
                 }
-                expect(sessionStore.retrieveSession()).toEqual(expectedSession)
+                expect(sessionStoreStrategy.retrieveSession()).toEqual(expectedSession)
                 expect(afterSession).toEqual(expectedSession)
                 done()
               },
             },
-            sessionStore
+            sessionStoreStrategy
           )
         })
       })
@@ -230,11 +230,11 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
       it('should abort after a max number of retry', () => {
         const clock = mockClock()
 
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
         stubStorage.setSpy.calls.reset()
 
         stubStorage.getSpy.and.returnValue(buildSessionString({ ...initialSession, lock: 'locked' }))
-        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStore)
+        processSessionStoreOperations({ process: processSpy, after: afterSpy }, sessionStoreStrategy)
 
         const lockMaxTries = isLockEnabled() ? LOCK_MAX_TRIES : 0
         const lockRetryDelay = isLockEnabled() ? LOCK_RETRY_DELAY : 0
@@ -254,14 +254,14 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
             retryState: initialSession,
           }),
         })
-        sessionStore.persistSession(initialSession)
+        sessionStoreStrategy.persistSession(initialSession)
 
         processSessionStoreOperations(
           {
             process: (session) => ({ ...session, value: 'foo' }),
             after: afterSpy,
           },
-          sessionStore
+          sessionStoreStrategy
         )
         processSessionStoreOperations(
           {
@@ -272,7 +272,7 @@ const initConfiguration: InitConfiguration = { clientToken: 'abc' }
               done()
             },
           },
-          sessionStore
+          sessionStoreStrategy
         )
       })
     })
