+++ packages/core/src/domain/session/storeStrategies/sessionStoreStrategy.ts
@@ -1,0 +1,7 @@
+import type { SessionState } from '../sessionState'
+
+export interface SessionStoreStrategy {
+  persistSession: (session: SessionState) => void
+  retrieveSession: () => SessionState
+  clearSession: () => void
+}
